compatible with long long ints

+++ general/reduce_fraction.c
@@ -2,9 +2,9 @@
 #include <stdio.h>
 #include "../main.h"
 
-int get_gcd(int a, int b)
+double get_gcd(long long int a, long long int b)
 {
-    int temp;
+    long long int temp;
     while (b != 0) {
         temp = b;
         b = a % b;
@@ -15,7 +15,7 @@ int get_gcd(int a, int b)
 
 void reduce_fraction(double *n, double *d)
 {
-    int gcd = get_gcd(*n, *d);
+    long long int gcd = get_gcd(*n, *d);
     (*n) /= gcd;
     (*d) /= gcd;
     printf("\nReduced fraction:\n%.2f/%.2f\n", *n, *d);
